// user-web/src/lib/types/api.generated.ts
export interface components {
    schemas: {
        CreateUserDto: {
            /** @description Nombre completo del usuario */
            name: string;
            /** @description Correo electrónico del usuario */
            email: string;
            /**
             * @description Rol jerárquico del usuario
             * @default client
             * @enum {string}
             */
            role: "client" | "admin" | "super_admin";
            /** @description Contraseña de acceso del usuario */
            password: string;
        };
        User: {
            /** @description ID autogenerado del usuario */
            id: number;
            /** @description Nombre completo del usuario */
            name: string;
            /** @description Correo electrónico del usuario */
            email: string;
            /**
             * @description Rol jerárquico del usuario
             * @default client
             * @enum {string}
             */
            role: "client" | "admin" | "super_admin";
        };
        UpdateUserDto: {
            /** @description Nombre completo del usuario */
            name?: string;
            /** @description Correo electrónico del usuario */
            email?: string;
            /**
             * @description Rol jerárquico del usuario
             * @default client
             * @enum {string}
             */
            role: "client" | "admin" | "super_admin";
            /** @description Contraseña de acceso del usuario */
            password?: string;
        };
        LoginDto: {
            /** @description Correo electrónico del usuario */
            email: string;
            /** @description Contraseña de acceso del usuario */
            password: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}